desk/pyproject.toml
[tool.poetry]
name = "desk"
version = "0.1.0"
description = ""
authors = ["Camila Arancibia <camilachillan99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.12.1"
pandas = "^2.0.3"
pyorbital = "^1.8.0"
dash-extensions = "^1.0.3"
python-dotenv = "^1.0.0"


[tool.poetry.group.format.dependencies]
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
